more structual changes so that flow path goes all down (now includes provenance data generation and the interface has been generalised for graph front end intergration)
